+++ src/app/components/UnratedPlayersList.tsx
@@ -5,6 +5,7 @@ import dynamic from 'next/dynamic';
 import { fetchGamePlayers } from "@/app/utils/playerDb";
 import { hasGameEnded } from "@/app/utils/gameUtils";
 import toast from "react-hot-toast";
+import { checkTimeSinceGameStarted } from "@/app/utils/gameUtils";
 
 // Dynamic import of PlayerItem with no SSR
 const PlayerItem = dynamic(() => import('./PlayerItem'), { 
@@ -55,6 +56,7 @@ export default function UnratedPlayersList({ playerId, gameId }: UnratedPlayersL
   const [gameLoading, setGameLoading] = useState<boolean>(true);
   // Add a new state to track if the user has already rated players in this game
   const [hasRatedGame, setHasRatedGame] = useState<boolean>(false);
+  const [isRatingWindowClosed, setIsRatingWindowClosed] = useState(false);
 
   const fetchPlayersAndRatings = async () => {
     if (!gameId || !playerId) return;
@@ -143,6 +145,21 @@ export default function UnratedPlayersList({ playerId, gameId }: UnratedPlayersL
     };
 
     fetchGame();
+  }, [gameId]);
+
+  useEffect(() => {
+    const checkRatingWindow = async () => {
+      if (!gameId) return;
+      
+      try {
+        const hoursSinceStart = await checkTimeSinceGameStarted(gameId);
+        setIsRatingWindowClosed(hoursSinceStart > 72); // 72 hours = 3 days
+      } catch (error) {
+        console.error('Error checking rating window:', error);
+      }
+    };
+
+    checkRatingWindow();
   }, [gameId]);
 
   // Function to store a rating in pending state
@@ -239,7 +256,8 @@ export default function UnratedPlayersList({ playerId, gameId }: UnratedPlayersL
   const isSubmitDisabled = isSubmitting || 
                           pendingRatings.length === 0 || 
                           isUserPlayer === false ||
-                          !isGameEnded;
+                          !isGameEnded ||
+                          isRatingWindowClosed;
 
   return (
     <div className="space-y-4">
@@ -254,6 +272,12 @@ export default function UnratedPlayersList({ playerId, gameId }: UnratedPlayersL
       {!isGameEnded && !gameLoading && (
         <div className="mb-4 p-3 rounded bg-red-600 text-white">
           This game hasn't ended yet. Ratings can only be submitted after the game has finished.
+        </div>
+      )}
+
+      {isRatingWindowClosed && (
+        <div className="mb-4 p-3 rounded bg-red-600 text-white">
+          The rating window for this game has closed. Ratings must be submitted within 72 hours after the game starts.
         </div>
       )}
       
@@ -349,9 +373,11 @@ export default function UnratedPlayersList({ playerId, gameId }: UnratedPlayersL
               ? 'Game Has Not Ended Yet'
               : isUserPlayer === false
                 ? 'Only Players Can Rate'
-                : pendingRatingsCount > 0
-                  ? `${hasRatedGame ? 'Update' : 'Submit'} ${pendingRatingsCount} Rating${pendingRatingsCount !== 1 ? 's' : ''}`
-                  : 'No Ratings to Submit'
+                : isRatingWindowClosed
+                  ? 'Rating Window Closed'
+                  : pendingRatingsCount > 0
+                    ? `${hasRatedGame ? 'Update' : 'Submit'} ${pendingRatingsCount} Rating${pendingRatingsCount !== 1 ? 's' : ''}`
+                    : 'No Ratings to Submit'
           }
         </button>
       </div>
